add scala-2.12 compatibility
